dilithium-signing-portal/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
liboqs-python
pydantic==2.5.0
python-multipart==0.0.6
